Refactor deploy into separate files
